requirements.txt
streamlit==0.89.0
requests==2.26.0
Pillow==9.0.1
transformers==4.12.2
torch==1.10.1
torchvision==0.11.3
